requirements.txt
flask
python-dotenv
scikit-learn
joblib
pathlib